requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
jinja2
pyyaml
email-validator
